fork-zipline/new_engine_starter/src/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
bcolz-zipline==1.13.0
blosc2==3.7.2
Bottleneck==1.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cramjam==2.11.0
empyrical-reloaded==0.5.12
exceptiongroup==1.3.0
exchange_calendars==4.11.1
fastapi==0.116.1
fastparquet==2024.11.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
h5py==3.14.0
idna==3.10
intervaltree==3.1.0
iso3166==2.1.1
iso4217==1.14.20250512
Jinja2==3.1.6
korean-lunar-calendar==0.3.1
lru-dict==1.3.0
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.1
multipledispatch==1.0.0
ndindex==1.10.0
networkx==3.4.2
numexpr==2.11.0
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pandas_market_calendars==5.1.1
patsy==1.0.1
peewee==3.17.3
platformdirs==4.4.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyluach==2.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.47.3
statsmodels==0.14.5
tables==3.10.1
tomli==2.2.1
toolz==1.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
zipline-reloaded==3.1.1
